refactor: global package 주소 수정
