+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/global/code/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.sk.skala.axcalibur.spec.global.code;
+package com.sk.skala.axcalibur.global.code;
 
 import org.springframework.http.HttpStatus;
