+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/global/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.sk.skala.axcalibur.spec.global.config;
+package com.sk.skala.axcalibur.global.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
